요약, 감정 모델 수정

+++ app/src/main/java/com/example/note/presentation/create/model/NoteModel.kt
@@ -9,6 +9,6 @@ data class NoteModel(
     val title: String,
     val content: String,
     val summarize: String,
-    val sentiment: Float,
+    val sentiment: Double,
     val createdDate: String
 ): Parcelable
